added spring security

+++ src/main/java/com/school/system/absence/AbsenceService.java
@@ -1,6 +1,14 @@
 package com.school.system.absence;
 
 import com.school.system.exception.NotFoundException;
+import com.school.system.policy.CreatePolicy;
+import com.school.system.policy.DeletePolicy;
+import com.school.system.policy.ReadPolicy;
+import com.school.system.policy.UpdatePolicy;
+import com.school.system.policy.exception.CannotCreateException;
+import com.school.system.policy.exception.CannotDeleteException;
+import com.school.system.policy.exception.CannotReadException;
+import com.school.system.policy.exception.CannotUpdateException;
 import com.school.system.subject.Subject;
 import com.school.system.subject.SubjectRepository;
 import com.school.system.users.parents.Parent;
@@ -8,6 +16,7 @@ import com.school.system.users.student.Student;
 import com.school.system.users.student.StudentRepository;
 import com.school.system.users.teacher.Teacher;
 import com.school.system.users.teacher.TeacherRepository;
+import com.school.system.users.user.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -36,7 +45,7 @@ public class AbsenceService {
         this.subjectRepository = subjectRepository;
     }
 
-    public AbsenceResponseDTO createAbsence(AbsenceRequestDTO absenceDTO) {
+    public AbsenceResponseDTO createAbsence(AbsenceRequestDTO absenceDTO, User user) {
         Absence toCreate = new Absence();
 
         Teacher teacher = teacherRepository.findById(absenceDTO.teacher())
@@ -51,6 +60,10 @@ public class AbsenceService {
         toCreate.setSubject(subject);
         toCreate.setCreatedAt(LocalDateTime.now());
 
+        if(!CreatePolicy.canCreateAbsence(user, toCreate)) {
+            throw new CannotCreateException();
+        }
+
         return AbsenceMapper.absenceToAbsenceResponseDTO(absenceRepository.save(toCreate));
     }
 
@@ -60,9 +73,24 @@ public class AbsenceService {
         return absences.map(absence -> AbsenceMapper.absenceToAbsenceResponseDTO(absence));
     }
 
-    public AbsenceResponseDTO updateAbsence(UUID id, AbsenceRequestDTO absenceDTO) {
+    public AbsenceResponseDTO getAbsence(UUID id, User user) {
+        Absence absence = absenceRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Absence not found"));
+
+        if(!ReadPolicy.canReadAbsence(user, absence)) {
+            throw new CannotReadException();
+        }
+
+        return AbsenceMapper.absenceToAbsenceResponseDTO(absence);
+    }
+
+    public AbsenceResponseDTO updateAbsence(UUID id, AbsenceRequestDTO absenceDTO, User user) {
         Absence toUpdate = absenceRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Absence not found"));
+
+        if(!UpdatePolicy.canUpdateAbsence(user, toUpdate)) {
+            throw new CannotUpdateException();
+        }
 
         if(toUpdate.getTeacher().getId() != absenceDTO.teacher()) {
             Teacher teacher = teacherRepository.findById(absenceDTO.teacher())
@@ -85,9 +113,14 @@ public class AbsenceService {
         return AbsenceMapper.absenceToAbsenceResponseDTO(absenceRepository.save(toUpdate));
     }
 
-    public void deleteAbsence(UUID id) {
+    public void deleteAbsence(UUID id, User user) {
         Absence toDelete = absenceRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Absence not found"));
+
+        if(!DeletePolicy.canDeleteAbsence(user, toDelete)) {
+            throw new CannotDeleteException();
+        }
+
         absenceRepository.delete(toDelete);
     }
 }
